+++ sistema-gerenciador-de-termos/adhesion.py
